requirements.txt
streamlit
llama-index
python-dotenv
llama-index-readers-wikipedia
